systems/modules: add a profile.externalbuilder…
… module.

It will have to move elsewhere (in modules) but has what is
required. It gets information from hosts.yml although the list of
buildMachines is currently static (ideally we would compute the list
from the `hosts.*` that have a `builder` attrSet).

It also filter the list of buildMachines if the machine is also the
current host. This is the case, for example, for aomi.home (which is a
builder as well as one that enable external builds).

Signed-off-by: Vincent Demeester <vincent@sbr.pm>