tools/emacs: detect project name from project.el
We were doing this for projectile but not for project.el. As I am
switching to this (and removing projectile), I needed the same kind of
support. The only trick is that project.el "project name" is a path,
and I didn't want to have it the complete path, so I am "skipping"
some prefix and using relative path for the rest.

Signed-off-by: Vincent Demeester <vincent@sbr.pm>