users/vincent: re-enable gmail for backup and "moving" mails
I am forwaring everything to my main account (except work mails). This
is more to have a local backup and be able to move mails from gmail to
my main setup easily (the one that are not forwarded).

Later, I will disable this one and *do it* manually on a backup server
instead.

Signed-off-by: Vincent Demeester <vincent@sbr.pm>