feat(projects): add GitHub PR checkout functionality
This adds the ability to checkout GitHub PRs directly from Emacs using the gh CLI.
The feature includes a new interactive command `checkout-github-pr` that fetches
and displays a list of open PRs for selection, then checks out the chosen PR
locally. The command is bound to C-x p G in the project keymap and added to the
project switch commands menu.

Signed-off-by: Vincent Demeester <vincent@sbr.pm>