tools/emacs: replace org-mru-clock with my own functions
This is a very basic, non-optimized function, but it works.
`vde/org-clock-in-any-heading` list all the heading so that I can then
clock-in any.

The list has the heading and the file name so that it's easier to
filter things (e.g. type "tekton release" to find all release related
tasks).

It probably needs some enhancement like filtering "DONE" or
"CANCELLED" keywords.

Signed-off-by: Vincent Demeester <vincent@sbr.pm>