keyboards: bootstrap moonlander configuration...
Going from the default keymap from upstream qmk. I also imported
source from oryx and have chmouel's configuration on the side to take
inspiration from it.

Just like what I did with zmk-config, this now compile but it is as
simple as it gets. Now I need to migrate the zmk-config keymap here
and add matrix led as well to help me.

Signed-off-by: Vincent Demeester <vincent@sbr.pm>