dots: add a folder for dotfiles not managed by home-manager
For dotfiles that change often or that need to be set also on other
systems that do not have NixOS or home-manager, this folder is there
for it.

Essentially, the idea is to be able to quickly edit and update some of
my dotfiles without having to rebuild the full system.

Signed-off-by: Vincent Demeester <vincent@sbr.pm>